% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_series.r
\name{aggregate_eun}
\alias{aggregate_eun}
\title{Aggregate Aglink results from E15 and NMS to EUN
Reason: sometimes the totals are missing.}
\usage{
aggregate_eun(datacube, attrib, product)
}
\arguments{
\item{datacube}{Aglink dataset as prepared by data-raw/aglink_timeseries.r}

\item{attrib}{Aglink attributes (e.g. DEL: deliveries)}

\item{product}{Aglink product code (e.g. MK: milk)}
}
\description{
Aggregate Aglink results from E15 and NMS to EUN
Reason: sometimes the totals are missing.
}
