% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_series.r
\name{aggregate_eun_list}
\alias{aggregate_eun_list}
\title{Aggregate Aglink results from E15 and NMS to EUN
This function expects lists of attributes and products
Reason: sometimes the totals are missing.}
\usage{
aggregate_eun_list(datacube, attrib_list, product_list)
}
\arguments{
\item{datacube}{Aglink dataset as prepared by data-raw/aglink_timeseries.r}

\item{attrib_list}{Aglink attributes (e.g. DEL: deliveries)}

\item{product_list}{Aglink product codes (e.g. MK: milk)}
}
\description{
Aggregate Aglink results from E15 and NMS to EUN
This function expects lists of attributes and products
Reason: sometimes the totals are missing.
}
