% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capri_gdx_util.r
\name{capri_filter}
\alias{capri_filter}
\title{Load CAPRI results and filter to a user-specified subset of results}
\usage{
capri_filter(
  filename,
  selregion = "all",
  seldim5 = "",
  selcols,
  selrows,
  simyear = "2030",
  scenarioname = "baseline"
)
}
\arguments{
\item{filename}{Name of the file}

\item{selregion}{Character vector of regions, default = "all"}

\item{seldim5}{Selection of the elements in the fifth dimension of the CAPRI data cube. By default it is the empty element}

\item{selcols}{Selection of columns in the CAPRI data cube}

\item{selrows}{Selection of rows in the CAPRI data cube}

\item{simyear}{Simulation year to be loaded/filtered}

\item{scenarionam}{Name of the scenario}
}
\description{
Load CAPRI results and filter to a user-specified subset of results
}
