% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables_detailed.r
\name{convert_balance_detailed}
\alias{convert_balance_detailed}
\title{Get detailed balance tables (demand broken down to its comoponents)}
\usage{
convert_balance_detailed(
  region_list,
  product_list,
  scenario_list,
  folder = "mydata"
)
}
\arguments{
\item{region_list}{List of regions (CAPRI code) for which the market balances should be derived}

\item{product_list}{List of commodities (CAPRI code) for which the market balances should be derived}

\item{scenario_list}{List of CAPRI scenarios for which the market balances should be derived}

\item{folder}{Path to folder containing the CAPRI result files}
}
\value{

}
\description{
Get detailed balance tables (demand broken down to its comoponents)
}
\examples{

# CAPRI BALANCES (NO INTRA-TRADE)
#-------------------------------

# define regions and commodities -- for which products do you need the market balances?
load("data/eu_region.RData")

meat_product       <- c("PORK", "POUM", "BEEF", "SGMT")
dairy_product      <- c("MILK", "BUTT", "CREM", "FRMI", "CHES", "SMIP", "COCM", "WMIO", "CASE", "WHEP")
cereal_product     <- c("CERE", "RYEM", "WHEA", "OATS", "BARL", "OCER", "MAIZ", "RICE")
oilseeds_product   <- c("RAPE", "SOYA", "SUNF")
cakes_product      <- c("RAPC", "SUNC", "SOYC", "CAKS")
oils_product       <- c("RAPO", "SUNO", "SOYO", "OLIO", "PLMO")

baseline_scenarios <- c("res_2_0810mtr_rd_ref", "res_2_0813mtr_rd_ref", "res_2_0820mtr_rd_ref", "res_2_0825mtr_rd_ref", "res_2_0830mtr_rd_ref")

# get market balances

meat_balance       <- convert_balance_detailed(eu_region, meat_product, baseline_scenarios, folder = "mydata")
cereal_balance     <- convert_balance_detailed(eu_region, cereal_product, baseline_scenarios, folder = "mydata")
dairy_balance      <- convert_balance_detailed(eu_region, dairy_product, baseline_scenarios, folder = "mydata")
oilseeds_balance   <- convert_balance_detailed(eu_region, oilseeds_product, baseline_scenarios, folder = "mydata")
cakes_balance      <- convert_balance_detailed(eu_region, cakes_product, baseline_scenarios, folder = "mydata")
oils_balance       <- convert_balance_detailed(eu_region, oils_product, baseline_scenarios, folder = "mydata")
sugar              <- convert_balance_detailed(eu_region, c("SUGA"), baseline_scenarios, folder = "mydata")

}
