% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.r
\name{convert_balance_ntrd}
\alias{convert_balance_ntrd}
\title{Convert market balances into a pre-defined format (for reporting purposes)}
\usage{
convert_balance_ntrd(
  region_list,
  product_list,
  scenario_list,
  folder = "mydata"
)
}
\arguments{
\item{region_list}{A character list of regions}

\item{product_list}{List of commodities}

\item{scenario_list}{List of scenarios}

\item{folder}{Path to folder with result files, default "mydata"}
}
\value{
A tibble with the reporting table
}
\description{
1: calcualate nettrade
2: append years (2010, 2013, 2020, 2025, 2030)
}
