% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables_detailed.r
\name{convert_product_balance}
\alias{convert_product_balance}
\title{Get the product balances for all baseline years and all products}
\usage{
convert_product_balance(
  region_list,
  product_list,
  scenario_list,
  folder = "mydata"
)
}
\arguments{
\item{region_list}{List of regions (CAPRI code) for which balances should be derived}

\item{product_list}{List of commodities (CAPRI code) for which balances should be derived}

\item{scenario_list}{List of CAPRI scenarios for which  balances should be derived}

\item{folder}{Path to folder containing the CAPRI result files}
}
\value{
A tibble with product balances
}
\description{
Get the product balances for all baseline years and all products
}
