% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables_detailed.r
\name{convert_supply_details}
\alias{convert_supply_details}
\title{Get the Farm|Supply details tables for all regions/activities/year}
\usage{
convert_supply_details(
  region_list,
  product_list,
  scenario_list,
  folder = "mydata"
)
}
\arguments{
\item{region_list}{List of regions (CAPRI code) for which product data should be derived}

\item{product_list}{List of commodities (CAPRI code) for which product data should be derived}

\item{scenario_list}{List of CAPRI scenarios for which  product datashould be derived}

\item{folder}{Path to folder containing the CAPRI result files}
}
\description{
Get the Farm|Supply details tables for all regions/activities/year
}
