% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_GUI_tables.r
\name{extract_gui_table}
\alias{extract_gui_table}
\title{Extracts pre-defined thematic tables from the results data cube}
\usage{
extract_gui_table(
  region_list,
  dim5_list,
  cols_list,
  rows_list,
  scenario_list,
  folder = "mydata"
)
}
\arguments{
\item{region_list}{A character list of regions}

\item{datacube}{A dplyr table with the raw capmod results}

\item{product_list}{List of commodities}

\item{scenario}{Scenario for which you want to retrieve results}
}
\value{
A dplyr table (tibble) containing the market balance
}
\description{
Extracts pre-defined thematic tables from the results data cube
}
