% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_GUI_tables.r
\name{filter_results_cube}
\alias{filter_results_cube}
\title{Generic function which filters the data cube}
\usage{
filter_results_cube(
  datacube,
  region_list,
  dim5_list,
  cols_list,
  rows_list,
  scenario_name
)
}
\arguments{
\item{datacube}{R object with full CAPRI resutls}

\item{region_list}{List of Regions to narrowed down on}

\item{dim5_list}{List of the fifth dimension elements}

\item{cols_list}{List of the elements in the column (COLS)}

\item{rows_list}{List of elements in the rows (ROWs)}

\item{scenario_name}{Name of the scenario you wish}
}
\value{
tibble with filtered results
}
\description{
Generic function which filters the data cube
}
