% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gdxfiles.r
\name{get_capmod_res}
\alias{get_capmod_res}
\title{Convert original CAPRI results into R format: gdx --> tbl_df --> RData
Results are only saved to the out_folder but not loaded into memory
You have to load() the RData file first before you can use it in R, but
this is usually done by the extraction scripts and not manually. See e.g. extract_gui_table().}
\usage{
get_capmod_res(
  scenario,
  in_folder,
  out_folder,
  gamspath = "/opt/GAMS",
  autoload = FALSE
)
}
\arguments{
\item{scenario}{Scenario file name. The usual structure applies:
res_ + regional break-down(0,2,999) + _ + baseyear + simulation year}

\item{in_folder}{Folder containing result file. Usually 'results/capmod' in the CAPRI installation}

\item{out_folder}{Folder for the converted formats}

\item{gamspath}{Path to gams installation. GAMS libraries are needed for the gdxrrw package}
}
\description{
Convert original CAPRI results into R format: gdx --> tbl_df --> RData
Results are only saved to the out_folder but not loaded into memory
You have to load() the RData file first before you can use it in R, but
this is usually done by the extraction scripts and not manually. See e.g. extract_gui_table().
}
