% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.r
\name{get_cowmilk}
\alias{get_cowmilk}
\title{Extracts cow milk production from CAPRI results}
\usage{
get_cowmilk(region_list, year_list)
}
\arguments{
\item{region_list}{list of regions}

\item{year_list}{list of years}
}
\value{
A tibble with cow milk supply results
}
\description{
Extracts cow milk production from CAPRI results
}
