% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.r
\name{get_cowmilk_aux}
\alias{get_cowmilk_aux}
\title{Auxiliary function for cow milk reporting}
\usage{
get_cowmilk_aux(
  datacube,
  region_list,
  attrib3_list = c("GROF", "PRCC", "DCOW"),
  year
)
}
\description{
Auxiliary function for cow milk reporting
}
