% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.r
\name{get_dairy}
\alias{get_dairy}
\title{Gets specific results on dairying}
\usage{
get_dairy(
  region_list,
  product_list = c("DCOH", "DCOL"),
  scenario_list,
  folder = "mydata"
)
}
\arguments{
\item{region_list}{List of regions}

\item{folder}{Folder where the baseline .RData files are stored. Default "mydata"}

\item{year_list}{List of simulation years}
}
\value{
tibble with dairy results
}
\description{
Gets specific results on dairying
}
