% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.r
\name{get_dairy_aux}
\alias{get_dairy_aux}
\title{Auxiliary function to load results (LEVL, YILD) related to dairying.
Both low and high-intensity variants included}
\usage{
get_dairy_aux(datacube, region_list, product_list = c("DCOH", "DCOL"))
}
\description{
Auxiliary function to load results (LEVL, YILD) related to dairying.
Both low and high-intensity variants included
}
