% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_GUI_tables.r
\name{get_GUI_table}
\alias{get_GUI_table}
\title{GEts pre-defined thematic tables direclty from a result folder}
\usage{
get_GUI_table(table = "supply details", scenario_list, folder = "mydata")
}
\arguments{
\item{datacube}{A dplyr table with the raw capmod results}

\item{region_list}{A character list of regions}

\item{product_list}{List of commodities}

\item{scenario}{Scenario for which you want to retrieve results}
}
\value{
A dplyr table (tibble) containing the market balance
}
\description{
GEts pre-defined thematic tables direclty from a result folder
}
\examples{

my_scenarios <- c("res_2_0830ghg_refpol_endotech_set12")
supply_table <- get_GUI_table(table = "supply details", my_scenarios, folder = "mydata")

}
