% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.r
\name{get_market_balance}
\alias{get_market_balance}
\title{Get market balances without intra trade from CAPMOD results}
\usage{
get_market_balance(datacube, region_list, product_list)
}
\arguments{
\item{datacube}{A dplyr table with the raw capmod results}

\item{region_list}{A character list of regions}

\item{product_list}{List of commodities}
}
\value{
A tibble containing the market balance
}
\description{
Get market balances without intra trade from CAPMOD results
}
