% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.r
\name{get_milk_deliveries}
\alias{get_milk_deliveries}
\title{Auxiliary function: extracts milk deliveries}
\usage{
get_milk_deliveries(datacube, region_list, product_list = c("PRCC"))
}
\description{
Auxiliary function: extracts milk deliveries
}
