% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.r
\name{get_product_balance}
\alias{get_product_balance}
\title{Get product balances from CAPMOD results}
\usage{
get_product_balance(datacube, region_list, product_list)
}
\arguments{
\item{datacube}{A dplyr table with the raw capmod results}

\item{region_list}{A character list of regions}

\item{product_list}{List of commodities}
}
\value{
A tibble containing the product balance
}
\description{
Get product balances from CAPMOD results
}
