% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.r
\name{get_supply_detail}
\alias{get_supply_detail}
\title{Get the table Farm Supply details}
\usage{
get_supply_detail(datacube, region_list, product_list)
}
\arguments{
\item{datacube}{A dplyr table with the raw capmod results}

\item{region_list}{A character list of regions}

\item{product_list}{List of commodities}
}
\value{
A tibble containing the supply details
}
\description{
Get the table Farm Supply details
}
