% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_series.r
\name{get_time_serie}
\alias{get_time_serie}
\title{Get Aglink time series from the _Ori.gdx file}
\usage{
get_time_serie(datacube, region_list, attrib1_list, attrib2_list)
}
\arguments{
\item{datacube}{A dplyr table with the raw capmod results}

\item{region_list}{A list of regions}

\item{attrib1_list}{A character vector of first attributes (usually balance items)}

\item{attrib2_list}{A character vector of second attributes (usually commodities)}
}
\value{
A dply table containing time series (all available years)
}
\description{
Get Aglink time series from the _Ori.gdx file
}
