% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capri_gdx_util.r
\name{load_dataout}
\alias{load_dataout}
\title{CAPRI gdx utilities}
\usage{
load_dataout(filename)
}
\arguments{
\item{filename}{Name of the .gdx file to be loaded}
}
\description{
simply loads the data cube from a CAPMOD result file
}
\examples{
load_dataout("test.gdx")

}
