% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capri_gdx_util.r
\name{pchange}
\alias{pchange}
\title{function to calculate percentage changes}
\usage{
pchange(a, b)
}
\arguments{
\item{a}{Base number of the percentage change calculation}

\item{b}{Target number of the percentage change calculation}
}
\description{
function to calculate percentage changes
}
