% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capri_gdx_util.r
\name{write_param_togdx}
\alias{write_param_togdx}
\title{write a parameter into a .gdx file}
\usage{
write_param_togdx(x, file, symname = "default", ts = "default")
}
\arguments{
\item{x}{R object}

\item{file}{Name of the output .gdx}

\item{symname}{Name of the GAMS parameter in the output .gdx file}

\item{ts}{GAMS parameter description}
}
\description{
write a parameter into a .gdx file
}
